refactor(ui-kit): 모듈 트리 변경
