+++ ui-kit/.storybook/preview.js
@@ -1,6 +1,6 @@
 import '../src/sass/index.scss';
 import React from 'react';
-import { LubyconUIKitProvider, Container } from '../src/components';
+import { LubyconUIKitProvider, Container } from '../src/index';
 
 export const decorators = [(Story => (
   <LubyconUIKitProvider>
