--- ui-kit/src/components/index.ts
@@ -1,15 +1,0 @@
-export { default as Alert } from './Alert';
-export { default as Button } from './Button';
-export { default as Checkbox } from './Checkbox';
-export { Row, Column } from './Grid';
-export { default as Container } from './Container';
-export { default as Radio } from './Radio';
-export { default as Selection } from './Selection';
-export { default as Switch } from './Switch';
-export { default as Text } from './Text';
-export { default as LubyconUIKitProvider } from './LubyconUIKitProvider';
-export { default as Toast } from './Toast';
-export { default as Tooltip } from './Tooltip';
-export { Tabs, TabPane } from './Tabs';
-export { default as Card, CardHeader, CardContent, CardImageContent, CardFooter } from './Card';
-export { default as Snackbar } from './Snackbar';
